requirements.txt

<old_str>flask==3.1.1
psycopg2-binary==2.9.10
python-telegram-bot==20.8
requests==2.32.3
sqlalchemy==2.0.41
python-dotenv==1.0.1</old_str>
<new_str>flask==3.1.1
psycopg2-binary==2.9.10
python-telegram-bot==20.8
requests==2.32.3
sqlalchemy==2.0.41
python-dotenv==1.0.1
urllib3>=1.26.0</new_str>
